Update: flie changes, added table q&a
Vse poizvedbe ki si bile prej v user.py models so sedaj v svojih file-h.
